Fix stretchy tracks
